refactor(analysis_artifact): rename analysis_format package
